% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{clean_eload}
\alias{clean_eload}
\title{Clean the elaod data}
\usage{
clean_eload(Data, n = 0.2, L_ptresh = 0.005, U_ptresh = 0.995)
}
\arguments{
\item{Data}{A dataframe of training or prediction data.}

\item{n}{An integer that correspond to a multiplicative coefficient that is used
to define the thresholding value. The default value is 0.2 which correspond to \emph{20} percent.}

\item{L_ptresh}{A numeric that correspond to the probability of the lower quantile used for filtering}

\item{U_ptresh}{A numeric that correspond to the probability of the upper quantile used for filtering}
}
\value{
A dataframe that correspond to the cleaned data
}
\description{
\code{clean_eload} This function remove the observations that have negative eload values or
values higher or lower than some upper and lower thresholds. The upper threshold is defined as
\emph{n} percent higher than the quantile corresponding to the given upper probability \emph{U_ptresh} and the lower
 threshold is defined as \emph{n} percent lower than the quantile corresponding to the given lower probability \emph{L_ptresh}
 divided by \emph{n}.
}

