% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{gbm_fit_plot}
\alias{gbm_fit_plot}
\title{Plot training period data}
\usage{
gbm_fit_plot(gbm_baseline_obj)
}
\arguments{
\item{gbm_baseline_obj}{A gbm_baseline_obj}
}
\value{
A dygraph time series plot
}
\description{
\code{gbm_fit_plot} Read a gbm_baseline object and plots the actual and the fitted data
}
