% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{gbm_plot}
\alias{gbm_plot}
\title{Plot training and prediction periods data}
\usage{
gbm_plot(gbm_baseline_obj, title = NULL)
}
\arguments{
\item{gbm_baseline_obj}{A gbm_baseline object}
}
\value{
A dygraph time series plot
}
\description{
\code{gbm_plot} Read a gbm_baseline object and plots the actual, the fitted and the predicted data
}

