% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{gbm_pred_plot}
\alias{gbm_pred_plot}
\title{Plot prediction period data}
\usage{
gbm_pred_plot(gbm_baseline_obj, title = NULL)
}
\arguments{
\item{gbm_baseline_obj}{A gbm_baseline object}
}
\value{
A dygraph time series plot
}
\description{
\code{gbm_pred_plot} Read a gbm_baseline object and plots the actual and the predicted data
}

