% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{pred_accuracy}
\alias{pred_accuracy}
\title{Prediction accuracy metrics computation}
\usage{
pred_accuracy(gbm_baseline_obj)
}
\arguments{
\item{gbm_baseline_obj}{A gbm_baseline object}
}
\value{
A dataframe with the computer R2, CV(RMSE) and NMBE
}
\description{
\code{pred_accuracy} This function compute the following prediction accuracy metrics:  R2, CV(RMSE) and NBME
}
