% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{time_format}
\alias{time_format}
\title{Convert the timestapms into the default format}
\usage{
time_format(data, format)
}
\arguments{
\item{data}{A dataframe that contains time column}

\item{format}{A character string that define the actual format of the timestamps.
Use the description of the base R function \emph{strptime} to define the format.}
}
\value{
A dataframe with timestamps converted into "\%m/\%d/\%y \%H:\%M" format
}
\description{
\code{time_format} This function convert the actual timestamps format into "\%m/\%d/\%y \%H:\%M" format
}

