% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{time_var}
\alias{time_var}
\title{Create new variables based on the timestamps.}
\usage{
time_var(Data)
}
\arguments{
\item{Data}{A dataframe of training or prediction data.}

\item{k_folds}{An integer that corresponds to the number of CV folds.}
}
\value{
A dataframe of training or prediction data including the new time variable.
}
\description{
\code{time_variables} splits the data into k folds by randomly selecting blocks of data,
where each block correspond to a calendar day.
}

